Move systems handler to handler package
